backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Starter template for Python backend projects"
authors = ["NRIDS"]
readme = "README.md"
#packages = [{ include = "backend_python" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.21"
pydantic = {extras = ["email"], version = "^2.5.2"}
requests = "^2.31.0"
psycopg2 = "^2.9.1"
pydantic-settings = "^2.1.0"
sqlmodel = "^0.0.14"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
ruff = "^0.3.4"

[tool.poetry.dev-dependencies]
prospector = "^1.10.3"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-watch = "^4.2.0"
alembic = "^1.13.1"
black = "^24.3.0"
isort = "^5.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
fail_under = 80

[tool.prospector]
strictness = "veryhigh"
test-warnings = true
doc-warnings = true

[tool.pytest.ini_options]
pythonpath = [".", "src"]
